requirements.txt
marker-pdf
pdftext
torch
transformers
pdf2image
uvicorn
fastapi
uvicorn
python-multipart
openai
google-genai
boto3
groq
pydantic-settings
eval_type_backport
PyYAML
python-dotenv
fhir.resources